backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.27.2
pyjwt==2.10.1
cryptography==43.0.3
openai==1.55.3
python-dotenv==1.0.1
